requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
pytest>=7.0.0
pandas>=1.3.0
fastparquet>=0.7.1
seaborne>=0.11.2
sklearn>=0.24.0
matplotlib>=3.4.0
scikit-learn>=0.24.0